Recipe builder API/requirements.txt
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
Flask-SQLAlchemy==2.5.1

